feat: Add support for capital ships, firing arcs and weapon bays. (#11)
* feat: Start working on targeting groups and weapon bays.

* feat: Mid-work commit.

* feat: Do necessary changes for UI.

* chore: Update nuget versions.

* fix: SpecialFeature None no longer exists.

* fix: Data fixes.

* fix: Data fixes.

* fix: Data errors.

* fix: Missing repository.

* fix: Less validation, filling takes care of this.

* fix: Weapon validation.

* chore: Update nuget versions.

* feat: Start working on better weapon dragdrop.

* chore: Update nuget versions.

* chore: Update nuget versions.

* fix: refresh.sh sample.

* fix: Refresh.sh removal because of propietary information.

* fix: Gitignore.

* gitignore

* fix: docker-compose.yml

* fix: Configuration.

* docker-compose

* feat: Cleanup eventing.

* chore: Fuck everything up.

* flail around: Broke it again, no idea how.

* feat: Continue working.

* fix: Simplify visual style.

* fix: Hiding condition.

* fix: Commit to not lose work.

* fix: Multiple fixes.
- ACs to 4/8/11/20
- Ammo usage fixes for weapon bays
- Heat usage fixes

* fix: Fixes after merge.

* fix: Allow weapon combination for capships.

* chore: Update nuget versions.

* fix: Arc diagrams were wrong.

* fix: Let's not leak connection strings to logs.

* fix: Evasion modifiers were not calculated correctly.

* fix: pushnugets.bat

* feat: Start removing https.

* flail around: Forwarding middleware.

* fix: Startup.cs

* fix: Finalize https -> http

* fix: Restore orignal port.

* feat: Add capital aerospace vessel support and a damage paperdoll for aerodyne dropships.

* chore: Update nuget versions.

* fix: Drophsip

* fix: rearrear

* fix: Front -> Structure for mechs/buildings.

* fix: Buildings do not receive through-armor criticals.
Capship picked wrong locations for damage.

* fix: Tooltip damage fetching for capships.

* fix: Self-review fixes.

* fix: Oops.

* flailing around: Try to address disconnection issue.

* flail: Revert Index.razor

* fix: Minor leave game issues.

* fix: Clarify logging, reorder.

* fix: Also ask for damage reports when joining a game.

* fix: Do not send null data.

* fix: Logging.

* fix: Do not attempt to display damage reports if game state is not present.

* fix: Multiple bugs.

* fix: Improve performance - Make a separate state for damage reports.

* fix: Bugsukka.

* fix: Serialization issue for unit lists in player states.

* fix: Experiment.

* feat: Make an exporter.
- Fix some stuff with targeting
- Add default values to Unit/UnitEntry
- Move Gunnery/Piloting to UnitEntry

* fix: Unit data screen still had gunnery/piloting.

* fix: Flail around.

* fix: Fix a lot of bugs.

* fix: Remove vulnerability with debug settings.

* fix: Try to survive lack of debug settings.

* fix: Revert fuckup.

* fix: Refresh issues, hub/communicator dispose issues.

* fix: Erraneous units are now CLEARLY visible.

* fix: Changelog.

* fix: Dispose manually.

* chore: New LZMA version.

* chore: Cleanup.

* feat: Separate ImprovedSensors from BAP (no logic yet).

* feat: Implement weapon individual hit modifier and BAP from TacOps.

* fix: Hit modifier calculation now needs WeaponEntry.

* fix: Visual style for amount/modifier in weaponentry.

* feat: Styling.

* fix: Undo sibling tag bug.

* fix: Padding.

* feat: Show modifier in weapon entry.

* fix: Re-export.

* fix: Restrict infantry movement modes further.

* fix: Sinks -> Dissipation

* fix: Sinks -> Dissipation

* feat: More mechs.

* feat: Automate sdk creation.

* juttuja

* fix: Continue to automate.

* chore: flail around

* chore: More flailing.

* fix: Flail more 2.

* chore: Expand.

* feat: Spectating.

* fix: Bugs.

* fix: Spectator list.

* fix: Spectator should see damage reports always.

* fix: Spectator list.

* refresh

* refresh

* fix: Cleanup.

* chore: Flail around, simplify.

* chore: More flailing.

* chore: More flail.

* fix: OnUnitListUpdated.

* fix: Do not redraw all unit entries on unit list change.

* fix: Remove key.

* chore: Get stupid.

* chore: Flail stop

* chore: Flail snail.

* fix: Revert.

* fix: Revert more.

* fix: Refresh needed when connecting.

* fix: Try to get rid of a deprecated nuget.

* fix: Remove deprecated nuget.

* fix: Null reference on invoke.

* fix: Commenting.

* fix: Move spectator control to tools.

* fix: Weapon display.

* fix: Only display weapon ammo selector if weapon has ammo.

---------

+++ CompressionLzma/src/CompressionLzma/CompressionHelper.cs
@@ -1,6 +1,5 @@
 // 7Zip helper code, original by Peter Bromberg
 // http://www.eggheadcafe.com/tutorials/aspnet/064b41e4-60bc-4d35-9136-368603bcc27a/7zip-lzma-inmemory-com.aspx
-using System;
 using System.IO;
 
 namespace SevenZip.Compression.LZMA;
@@ -35,12 +34,12 @@ public static class CompressionHelper
     // these are the default properties, keeping it simple for now:
     private static readonly object[] Properties =
     {
-        (Int32)Dictionary,
-        (Int32)2,
-        (Int32)3,
-        (Int32)0,
-        (Int32)2,
-        (Int32)128,
+        Dictionary,
+        2,
+        3,
+        0,
+        2,
+        128,
         "bt4",
         Eos
     };
